+++ sys/src/9/port/devmnt.c
@@ -47,7 +47,7 @@ struct Mntrpc
 
 enum
 {
-	TAGSHIFT = 5,			/* ulong has to be 32 bits */
+	TAGSHIFT = 5,			/* uint32_t has to be 32 bits */
 	TAGMASK = (1<<TAGSHIFT)-1,
 	NMASK = (64*1024)>>TAGSHIFT,
 };
@@ -61,7 +61,7 @@ struct Mntalloc
 	int	nrpcfree;
 	int	nrpcused;
 	uint	id;
-	ulong	tagmask[NMASK];
+	uint32_t	tagmask[NMASK];
 }mntalloc;
 
 Mnt*	mntchk(Chan*);
